refactor: edit dialog 분리
